consumer/requirements.txt
certifi==2025.8.3
elastic-transport==9.1.0
elasticsearch==9.1.0
kafka-python==2.2.15
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
